fix(discovery): resolve all mypy type errors and improve type safety
Comprehensive type safety improvements across the discovery module:

Type Errors Fixed (42 total across 15 files):
- Remove 24 unused type: ignore comments from API routers
- Fix SQLAlchemy execute parameter types in faceted_search.py
- Fix numpy float return type in search_ranking.py
- Fix numpy array loading in semantic_search.py
- Fix middleware dispatch callable type annotation
- Fix Redis ping await type with proper casting
- Fix Pool attribute access in db_pool_metrics.py
- Create properly typed rate limit handler wrapper

Implementation Details:
- Replace SQLAlchemy positional parameters with bindparam(expanding=True)
- Add explicit float() conversions for numpy operations
- Use cast() for redis-py and numpy type stub inconsistencies
- Use getattr() with defaults for SQLAlchemy async pool attributes
- Add proper Callable[[Request], Awaitable[Response]] type annotation

Configuration Updates (pyproject.toml):
- Add scipy.*, onnxruntime, optimum.* to mypy ignored imports
- Add module-specific overrides for untyped decorators
- Add overrides for files using untyped ML libraries

All mypy errors resolved, all pre-commit hooks passing.

+++ discovery/search_ranking.py
@@ -530,7 +530,7 @@ class SearchRanker:
             similarities.append(similarity)
 
         # Diversity is inverse of average similarity
-        avg_similarity = np.mean(similarities) if similarities else 0.0
+        avg_similarity = float(np.mean(similarities)) if similarities else 0.0
         diversity = 1.0 - avg_similarity
 
-        return max(0.0, min(diversity, 1.0))
+        return float(max(0.0, min(diversity, 1.0)))
